requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
coverage==7.8.0
execnet==2.1.1
Faker==37.1.0
greenlet==3.2.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
pytest==7.4.0
pytest-base-url==2.1.0
pytest-cov==6.1.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
pytest-reportlog==0.4.0
pytest-rerunfailures==15.0
pytest-sugar==1.0.0
pytest-xdist==3.6.1
python-dotenv==1.1.0
python-slugify==8.0.4
requests==2.32.5
termcolor==3.0.1
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.6.3
pre-commit==4.5.1
mypy==1.19.1
ruff==0.15.0
bandit[toml]==1.9.3
pip-audit==2.10.0
radon==6.0.1
hypothesis==6.151.5
mutmut==2.4.0
